:recycle: answer 컴포넌트 명 option으로 변경, 내부 비지니스 로직 test로 옮김
